pyproject.toml
[project]
name = "piutil"
version = "0.1.0"
description = "Lightweight profiling utilities for ML training pipelines"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
torch = ["torch>=2.0"]
tb = ["tensorboard>=2.14"]
data = ["webdataset>=0.2.86", "numpy>=1.24"]
dali = ["nvidia-dali-cuda120>=1.35"]
all = ["torch>=2.0", "tensorboard>=2.14", "webdataset>=0.2.86", "numpy>=1.24"]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/piutil"]

[tool.pytest.ini_options]
testpaths = ["tests"]
